pyproject.toml
[project]
name = "eth-wallet-risk-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.78.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
